提交文件

+++ cloud-consumer-order80/src/main/java/com/atguigu/springcloud/controller/OrderController.java
@@ -23,7 +23,7 @@ public class OrderController {
 //        return restTemplate.postForObject(PAYMENT_URL+"/payment/create",payment, CommonResult.class);
 //    }
     @GetMapping("/consumer/payment/get/{id}")
-    public CommonResult<Payment> getPaymentById(@PathVariable("id") Long id ){
+    public CommonResult<Payment> getPaymentById(@PathVariable("id") int id ){
         return orderFeignService.getPaymentById(id);
     }
     @PostMapping(value = "/payment/create")
