+++ cloud-consumer-order80/src/main/java/com/atguigu/springcloud/service/PaymentFeignService.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @FeignClient(value = "CLOUD-PAYMENT-SERVICE",fallback = PaymentFeignServiceImpl.class)
 public interface PaymentFeignService {
     @RequestMapping("/payment/get/{id}")
-    CommonResult getPaymentById(@PathVariable("id") Long id);
+    CommonResult getPaymentById(@PathVariable("id") int id);
 
     @PostMapping(value = "/payment/create")
     CommonResult create(@RequestBody Payment payment);
